+++ videosnap/VideoSnap.h
@@ -3,65 +3,69 @@
 //  VideoSnap
 //
 //  Created by Matthew Hutchinson on 18/08/2013.
-//  Copyright (c) 2013 Matthew Hutchinson. All rights reserved.
+//  Copyright (c) 2016 Matthew Hutchinson. All rights reserved.
 //
 
-#import <Cocoa/Cocoa.h>
-#import <QTKit/QTKit.h>
+#import <CoreMedia/CoreMedia.h>
+#import <AVFoundation/AVFoundation.h>
 
-// logging helpers
+
+#define VERSION @"0.0.2"
+
+// logging
 #define error(...) fprintf(stderr, __VA_ARGS__)
 #define console(...) printf(__VA_ARGS__)
 #define verbose(...) (is_verbose && fprintf(stderr, __VA_ARGS__))
-
-// version
-#define VERSION @"0.0.1"
+#define verbose_error(...) (is_verbose && fprintf(stderr, __VA_ARGS__))
 
 // defaults
+#define DEFAULT_FRAMES_PER_SECOND  @30
 #define DEFAULT_RECORDING_DELAY    @0.5
 #define DEFAULT_RECORDING_DURATION @6.0
 #define DEFAULT_RECORDING_FILENAME @"movie.mov"
-#define DEFAULT_RECORDING_SIZE     @"SD480"
-#define DEFAULT_VIDEO_SIZES        @[@"120", @"240", @"SD480", @"HD720"]
+#define DEFAULT_ENCODING_PRESET    @"Medium"
+#define DEFAULT_ENCODING_PRESETS   @[@"High", @"Medium", @"Low", @"640x480", @"1280x720"]
+
+// encoding preset options:
+// High - Highest recording quality (varies per device)
+// Medium - Suitable for WiFi sharing (actual values may change)
+// Low - Suitable for 3G sharing (actual values may change)
+// 640x480 - 640x480 VGA (check its supported before setting it)
+// 1280x720 - 1280x720 720p HD (check its supported before setting it)
 
 // default verbose flag
 BOOL is_verbose = NO;
 
 // VideoSnap
-@interface VideoSnap : NSObject {
-
-  QTCaptureSession         *captureSession;          // session
-  QTCaptureMovieFileOutput *captureMovieFileOutput;  // file output
-  QTCaptureDeviceInput     *captureVideoDeviceInput; // video input
-  QTCaptureDeviceInput     *captureAudioDeviceInput; // audio input
-  NSDate                   *recordingStartedDate;    // record timing
-  NSNumber                 *maxRecordingSeconds;     // record duration
+@interface VideoSnap : NSObject <AVCaptureFileOutputRecordingDelegate> {
+  AVCaptureSession         *session;
+  AVCaptureMovieFileOutput *movieFileOutput;
 }
+
 
 // class methods
 
 /**
- * Returns attached QTCaptureDevice objects that have video. Includes
- * video-only devices (QTMediaTypeVideo) and any audio/video devices
+ * Returns attached AVCaptureDevice objects that have video. Includes
+ * any devices supporting AVMediaTypeVideo and AVMediaTypeMuxed (audio & video)
  *
- * @return autoreleased array of video devices
+ * @return NSArray of video devices
  */
 +(NSArray *)videoDevices;
 
 /**
- * Returns default QTCaptureDevice object for video or nil if none found
+ * Returns the default AVCaptureDevice object for video or nil if none found
  *
- * @return QTCaptureDevice
+ * @return AVCaptureDevice
  */
-+(QTCaptureDevice *)defaultDevice;
++(AVCaptureDevice *)defaultDevice;
 
 /**
- * Returns QTCaptureDevice matching name or nil if a device matching the name
- * cannot be found
+ * Returns a AVCaptureDevice matching on localizedName or nil if not found
  *
- * @return QTCaptureDevice
+ * @return AVCaptureDevice
  */
-+(QTCaptureDevice *)deviceNamed:(NSString *)name;
++(AVCaptureDevice *)deviceNamed:(NSString *)name;
 
 /**
  * Captures video from a device and saves it to a file, returns (BOOL) YES if
@@ -69,7 +73,7 @@ BOOL is_verbose = NO;
  *
  * @return BOOL
  */
-+(BOOL)captureVideo:(QTCaptureDevice *)videoDevice filePath:(NSString *)path recordingDuration:(NSNumber *)recordSeconds videoSize:(NSString *)videoSize withDelay:(NSNumber *)delaySeconds noAudio:(BOOL)noAudio;
++(BOOL)captureVideo:(AVCaptureDevice *)videoDevice filePath:(NSString *)path recordingDuration:(NSNumber *)recordSeconds encodingPreset:(NSString *)encodingPreset delaySeconds:(NSNumber *)delaySeconds noAudio:(BOOL)noAudio;
 
 
 // instance methods
@@ -77,36 +81,26 @@ BOOL is_verbose = NO;
 -(id)init;
 
 /**
+ * Starts a capture session on a device for recordSeconds saving to filePath 
+ * using the encodingPreset (with an optional delay) returns (BOOL) YES if 
+ * successful
+ *
+ * @return BOOL
+ */
+-(BOOL)startSession:(AVCaptureDevice *)device filePath:(NSString *)path recordingDuration:(NSNumber *)recordSeconds encodingPreset:(NSString *)encodingPreset delaySeconds:(NSNumber *)delaySeconds noAudio:(BOOL)noAudio;
+
+/**
  * Adds an audio device to the capture session, uses the audio from videoDevice
  * if it is available, returns (BOOL) YES if successful
  *
  * @return BOOL
  */
--(BOOL)addAudioDevice:(QTCaptureDevice *)videoDevice;
+-(BOOL)addAudioDevice:(AVCaptureDevice *)videoDevice;
 
 /**
- * Sets compression video/audio options on the output file
+ * AVCaptureMovieFileOutput delegate, called when output file has been finally
+ * written to, due to error or stopping the capture session 
  */
--(void)setCompressionOptions:(NSString *)videoCompression audioCompression:(NSString *)audioCompression;
-
-/**
- * Starts a capture session on device, saving to a filePath for recordSeconds
- * return (BOOL) YES successful or (BOOL) NO if not
- *
- * @return BOOL
- */
--(BOOL)startSession:(QTCaptureDevice *)device filePath:(NSString *)path recordingDuration:(NSNumber *)recordSeconds videoSize:(NSString *)videoSize withDelay:(NSNumber *)withDelay noAudio:(BOOL)noAudio;
-
-/**
- * QTCaptureMovieFileOutput delegate called when camera samples from the output
- * buffer
- */
--(void)captureOutput:(QTCaptureFileOutput *)captureOutput didOutputSampleBuffer:(QTSampleBuffer *)sampleBuffer fromConnection:(QTCaptureConnection *)connection;
-
-/**
- * QTCaptureMovieFileOutput delegate, called when output file has been finally
- * written to
- */
--(void)captureOutput:(QTCaptureFileOutput *)captureOutput didFinishRecordingToOutputFileAtURL:(NSURL *)outputFileURL forConnections:(NSArray *)connections dueToError:(NSError *)error;
+-(void)captureOutput:(AVCaptureFileOutput *)captureOutput didFinishRecordingToOutputFileAtURL:(NSURL *)outputFileURL fromConnections:(NSArray *)connections error:(NSError *)error;
 
 @end