+++ CHANGELOG.md
@@ -5,12 +5,18 @@ project adheres to [Semantic Versioning][Semver].
 
 ## [Unreleased]
 
-  * WIP - refactor to use AV Foundation Framework
+  * WIP - allow capture to run indefinitely until ctrl+c pressed
+
+## [0.0.2][] (8 July 2016)
+  * Update to use the [AV Foundation](https://developer.apple.com/av-foundation/) framework
+  * Cocoa and QTKit frameworks dropped
+  * Encoding presets changed, use `-p` option to set (`-s` option dropped)
+  * New encoding presets are: High, Medium (default), Low, 640x480 or 1280x720
 
 ## [0.0.1][] (8 September 2013)
   * Initial release of VideoSnap (using the QTKit framework)
 
-[Unreleased]: https://github.com/matthutchinson/videosnap/compare/v0.0.1...HEAD
+[Unreleased]: https://github.com/matthutchinson/videosnap/compare/v0.0.2...HEAD
+[0.0.2]: https://github.com/matthutchinson/videosnap/compare/v0.0.1...v0.0.2
 [0.0.1]: https://github.com/matthutchinson/videosnap/releases/tag/v0.0.1
 [Semver]: http://semver.org
-